Add Wikimedia Foundation
From the announcement: "For the remainder of March 2020, our San Francisco office will be closed to staff and visitors. We are putting in place measures to ensure that our San Francisco-based staff have resources and support to continue working remotely. Our Washington D.C. office will remain open for the time being, though we are encouraging everyone to take precautions to protect themselves and their communities and to work remotely where possible.

We have also temporarily suspended nonessential travel for all staff, and instituted a risk review process for any travel considered essential. In addition, we have been in touch with members of the Wikimedia movement community with respect to upcoming events and are taking necessary steps to cancel or postpone these, based on potential risks."